최종 수익 int형태로 출력

+++ src/main/java/domain/user/Dealer.java
@@ -58,8 +58,8 @@ public class Dealer {
         cost -= playerCost;
     }
 
-    public double finalCost() {
-        return cost;
+    public int finalCost() {
+        return (int) cost;
     }
 
     public String toStringFirstTurn() {
